% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatureFromBED.R
\name{addFeatureFromBED}
\alias{addFeatureFromBED}
\title{Add feature to the model from a BED file}
\usage{
addFeatureFromBED(features, ranges, featureName = NULL, file = NULL)
}
\arguments{
\item{features}{a data.frame to add the features to}

\item{ranges}{a \code{GRanges} object with the ranges to overlap features
(must have same length as the number of rows in the features data frame)}

\item{featureName}{what the feature should be called}

\item{file}{path to the BED File with 4 columns (chr, start, end, feature)}
}
\value{
the original features data.frame with the new feature(s) added from
the BED file as the last column(s)
}
\description{
Helper function to add features from BED files. All features will become
binary features, 0 indicating no overlap, 1 indicating overlap.
If the feature has more than one factor, each factor will become its own
binary feature.
}
