% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFeatures.R
\name{constructFeatures}
\alias{constructFeatures}
\title{Feature construction}
\usage{
constructFeatures(bs, sample = 1, minCov = 10, sampleAvg = TRUE,
  neighbMeth = TRUE, neighbDist = TRUE, featureBEDs = NULL)
}
\arguments{
\item{bs}{a \code{BSseq} object}

\item{sample}{which sample to construct features for; default is 1, the
first sample.}

\item{minCov}{the minimum coverage, below which values will be set to NA.
Default is 10.}

\item{sampleAvg}{boolean of whether to not to include the sample average
as a feature. Default is TRUE.}

\item{neighbMeth}{boolean of whether or not to include nearest non-missing
neighboring CpG methylation values. Default is TRUE.}

\item{neighbDist}{boolean of whether or not to include nearest non-missing
neighboring CpG distances. Default is TRUE.}

\item{featureBEDs}{optional named list of paths to BED files to be included
as features in the model. Names are used as the feature name;
e.g. list(chromState = "chromatinStates.bed")}
}
\value{
a data.frame, each row is a CpG; first column is the actual beta
value, subsequent columns are features
}
\description{
Helper function to make a data frame ready for training/validation/testing
from a bsseq object that contains the CpG/DNA features for the model
}
