% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceMetrics.R
\name{performanceMetrics}
\alias{performanceMetrics}
\title{Performance metrics}
\usage{
performanceMetrics(pred, actual)
}
\arguments{
\item{pred}{a vector containing the predicted methylation proportions.}

\item{actual}{a vector containing the actual methylated proportions.}
}
\value{
a list of RMSE, AUROC, AUPRC, and accuracy
}
\description{
Helper function to calculate performance metrics from labelled predictions.
}
